use int instead of long for user id

+++ src/main/java/com/openclassrooms/chatop/model/User.java
@@ -18,7 +18,8 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+
+    private int id;
 
     private String email;
